+++ kikuchipy/imaging/vbse.py
@@ -16,7 +16,6 @@
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
 from typing import List, Optional, Tuple, Union
-import warnings
 
 from dask.array import Array
 from hyperspy.drawing._markers.horizontal_line import HorizontalLine
@@ -29,14 +28,14 @@ import numpy as np
 
 from kikuchipy.signals import EBSD, LazyEBSD
 from kikuchipy.signals import VirtualBSEImage
+from kikuchipy.pattern import rescale_intensity
 from kikuchipy._util._transfer_axes import _transfer_navigation_axes_to_signal_axes
-from kikuchipy.imaging.vbse import _get_rgb_image
 
 
-class VirtualBSEGenerator:
-    """*[Deprecated]* Generates virtual backscatter electron (BSE)
-    images for an EBSD signal and a set of EBSD detector areas in a
-    convenient manner.
+class VirtualBSEImager:
+    """Generate virtual backscatter electron (BSE) images for an
+    electron backscatter diffraction (EBSD) signal and a set
+    of EBSD detector areas in a convenient manner.
 
     Parameters
     ----------
@@ -47,28 +46,14 @@ class VirtualBSEGenerator:
     --------
     kikuchipy.signals.EBSD.plot_virtual_bse_intensity,
     kikuchipy.signals.EBSD.get_virtual_bse_intensity
-
-    Notes
-    -----
-    Deprecated since version 0.8: Class ``VirtualBSEGenerator`` is
-    deprecated and will be removed in version 0.9. Use
-    :class:`~kikuchipy.imaging.VBSEImager` instead.
     """
 
     def __init__(self, signal: Union[EBSD, LazyEBSD]):
         self.signal = signal
         self._grid_shape = (5, 5)
 
-        warnings.warn(
-            message=(
-                "Class `VirtualBSEGenerator` is deprecated and will be removed in "
-                "version 0.9. Use `kikuchipy.imaging.VirtualBSEImager` instead."
-            ),
-            category=np.VisibleDeprecationWarning,
-        )
-
     def __repr__(self):
-        return f"VirtualBSEGenerator for {self.signal}"
+        return self.__class__.__name__ + " for " + repr(self.signal)
 
     @property
     def grid_rows(self) -> np.ndarray:
@@ -254,8 +239,8 @@ class VirtualBSEGenerator:
 
     def roi_from_grid(self, index: Union[Tuple, List[Tuple]]) -> RectangularROI:
         """Return a rectangular region of interest (ROI) on the EBSD
-        detector from one or multiple generator grid tile indices as
-        row(s) and column(s).
+        detector from one or multiple grid tile indices as row(s) and
+        column(s).
 
         Parameters
         ----------
@@ -366,3 +351,116 @@ class VirtualBSEGenerator:
         pattern.add_marker(markers, permanent=True)
 
         return pattern
+
+
+def _normalize_image(
+    image: np.ndarray,
+    add_bright: int = 0,
+    contrast: float = 1.0,
+    dtype_out: Union[str, np.dtype, type] = "uint8",
+) -> np.ndarray:
+    """Normalize an image's intensities to a mean of 0 and a standard
+    deviation of 1, with the possibility to also scale by a contrast
+    factor and shift the brightness values.
+
+    Clips intensities to uint8 data type range, ``[0, 255]``.
+
+    Adapted from the aloe/xcdskd package.
+
+    Parameters
+    ----------
+    image
+        Image to normalize.
+    add_bright
+        Brightness offset to for each array. Default is ``0``.
+    contrast
+        Contrast factor for each array. Default is ``1.0``.
+    dtype_out
+        Output data type, either ``"uint8"`` (default) or ``"uint16"``.
+
+    Returns
+    -------
+    normalized_image
+        Normalized image.
+    """
+    dtype_out = np.dtype(dtype_out)
+    dtype_max = np.iinfo(dtype_out).max
+
+    offset = (dtype_max // 2) + add_bright
+    contrast *= dtype_max * 0.3125
+    median = np.median(image)
+    std = np.std(image)
+    normalized_image = offset + ((contrast * (image - median)) / std)
+
+    return np.clip(normalized_image, 0, dtype_max)
+
+
+def _get_rgb_image(
+    channels: List[np.ndarray],
+    percentiles: Optional[Tuple] = None,
+    normalize: bool = True,
+    alpha: Optional[np.ndarray] = None,
+    dtype_out: Union[str, np.dtype, type] = "uint8",
+    add_bright: int = 0,
+    contrast: float = 1.0,
+) -> np.ndarray:
+    """Return an RGB image from three numpy arrays, with a potential
+    alpha channel.
+
+    Parameters
+    ----------
+    channels
+        A list of np.ndarray for the red, green and blue channel,
+        respectively.
+    percentiles
+        Whether to apply contrast stretching with a given percentile
+        tuple with percentages, e.g. (0.5, 99.5), after creating the
+        RGB image. If not given (default), no contrast stretching is
+        performed.
+    normalize
+        Whether to normalize the individual ``channels`` before
+        RGB image creation. Default is ``True``.
+    alpha
+        Potential alpha channel. If not given (default), no alpha
+        channel is added to the image.
+    dtype_out
+        Output data type, either ``"uint8"`` (default) or ``"uint16"``.
+    add_bright
+        Brightness offset to for each array. Default is ``0``.
+    contrast
+        Contrast factor for each array. Default is ``1.0``.
+
+    Returns
+    -------
+    rgb_image
+        RGB image.
+    """
+    dtype_out = np.dtype(dtype_out)
+
+    n_channels = 3
+    rgb_image = np.zeros(channels[0].shape + (n_channels,), np.float32)
+    for i, channel in enumerate(channels):
+        if normalize:
+            channel = _normalize_image(
+                channel.astype(np.float32),
+                dtype_out=dtype_out,
+                add_bright=add_bright,
+                contrast=contrast,
+            )
+        rgb_image[..., i] = channel
+
+    # Apply alpha channel if desired
+    if alpha is not None:
+        alpha_min = np.nanmin(alpha)
+        rescaled_alpha = (alpha - alpha_min) / (np.nanmax(alpha) - alpha_min)
+        for i in range(n_channels):
+            rgb_image[..., i] *= rescaled_alpha
+
+    # Rescale to fit data type range
+    if percentiles is not None:
+        in_range = tuple(np.percentile(rgb_image, q=percentiles))
+    else:
+        in_range = None
+    rgb_image = rescale_intensity(rgb_image, in_range=in_range, dtype_out=dtype_out)
+
+    return rgb_image.astype(dtype_out)